move file state to domain
