+++ libs/design/FileUpload/index.ts
@@ -1,3 +1,2 @@
 export * from './ControlledFileUpload';
-export type { StorageFile } from './FileState';
 export * from './FileUpload';
