+++ libs/domain/FileState.ts
